==> USBR/Colorado River Basin Natural Flow and Salt Data/Summary Sheet/AggregatedAmounts/0_USBRSummaryAggregatedDataPreprocess.ipynb <==
## How to convert shapefile to WKT.

Need to Convert a shapefile to WKT so we can use it as a Geometry field in our WaDE imports.

# #Needed Libararies
# import os
# import numpy as np
# import pandas as pd
# import geopandas as gpd # the library that lets us read in shapefiles
# pd.set_option('display.max_columns', 999)  # How to display all columns of a Pandas DataFrame in Jupyter Notebook

# #Working Directory and Input Files
# workingDir = "C:/Users/rjame/Documents/WSWC Documents/MappingStatesDataToWaDE2.0/USBR/Colorado River Basin Natural Flow and Salt Data/Summary Sheet/AggregatedAmounts/RawInputData/byState"
# os.chdir(workingDir)

# USBR_Tri = gpd.read_file('C:/Users/rjame/Documents/WSWC Documents/MappingStatesDataToWaDE2.0/USBR/Colorado River Basin Natural Flow and Salt Data/Summary Sheet/AggregatedAmounts/RawInputData/byState/ByState.shp')

# USBR_Tri.head(20)

# df = pd.DataFrame(USBR_Tri)
# df

# # Printing file to xlsx
# df.to_excel('USBR_Tri.xlsx', index=False)  # The output

# Working with USBR Aggregated Data

Pre-processing input data for a smoother upload experience of the state data to the WaDE 2.0 database.
Using geopandas to read in shp file, and coverting to WKT for ReportingUnit geometry.

#### Notes:
- NA

#Needed Libararies
import os
import numpy as np
import pandas as pd
from datetime import datetime
import geopandas as gpd # the library that lets us read in shapefiles
pd.set_option('display.max_columns', 999)  # How to display all columns of a Pandas DataFrame in Jupyter Notebook

#Setting work directory, reading inputs, creating dataframe
workingDir = "C:/Users/rjame/Documents/WSWC Documents/MappingStatesDataToWaDE2.0/USBR/Colorado River Basin Natural Flow and Salt Data/Summary Sheet/AggregatedAmounts/RawInputData"
os.chdir(workingDir)
fileInput = "perfect example_A.csv"
df = pd.read_csv(fileInput)
df.head(3)

#Approach:
#Make a temp df to store in values from a list
#Append temp df to output df

target_columns = ["Amount", "BeneficialUseCV", "ReportYearCV", "STATE", "COUNTY", "WaterSource"]
dfout = pd.DataFrame(columns=target_columns)
dftemp = pd.DataFrame(columns=target_columns)

# for row in df.index:
for row in df.index:
    #Get Values
    Amounts = df.iloc[row].drop(["USE TYPE", "STATE", "COUNTY"]).values.tolist()
    BenUse = df.iloc[row,0]
    ReportYears = df.iloc[0].drop(["USE TYPE", "STATE", "COUNTY"]).index.values.tolist()
    State = str(df.iloc[row,1])
    
    CWS = str(df.iloc[row,2])
    if '-' in CWS:
        County = str(df.iloc[row,2].split('-')[0])
        WaterSource = str(df.iloc[row,2].split('-')[1])
    else:
        County = ""
        WaterSource = str(df.iloc[row,2])  
    
    #Put Values into temp Dataframe
    dftemp['Amount'] = Amounts
    dftemp['ReportYearCV'] = ReportYears
    dftemp['BeneficialUseCV'] = BenUse
    dftemp['STATE'] = State
    dftemp['COUNTY'] = County
    dftemp['WaterSource'] = WaterSource
    
    #Append to dfout
    dfout = dfout.append(dftemp)

dfout
dfout.reset_index()

# Convert State to abbreveated version
StateAbDic = {
"Arizona" : "AZ",
"Colorado" : "CO",
"New Mexico" : "NM",
"Utah" : "UT",
"Wyoming" : "WY"}

def assignStateAb(colrowValue):
    if colrowValue == '' or pd.isnull(colrowValue):
        outList = ""
    else:
        String1 = colrowValue  # remove whitespace chars
        try:
            outList = StateAbDic[String1]
        except:
            outList = ""
    return outList

dfout['STATE'] = dfout.apply(lambda row: assignStateAb(row['STATE']), axis=1)
dfout

# Creating ReportingUnitName field.

dfout['in_ReportingUnitName'] = dfout['STATE'] + "_" + dfout['WaterSource']
dfout

#Fixing Datatypes
dfout['Amount'] = dfout['Amount'].astype(float) 
dfout['ReportYearCV'] = dfout['ReportYearCV'].astype(int)
dfout.dtypes

# TimeframeStart & TimeframeEnd

dfout['in_TimeframeStart'] = '01/01/' + dfout['ReportYearCV'].astype(str)
dfout['in_TimeframeEnd'] = '12/31/' + dfout['ReportYearCV'].astype(str)
dfout.head(3)

## WaDE Custom Elements (due to missing reporting unit and water source info)

# Creating WaDE Custom water source native ID for easy water source identification
# ----------------------------------------------------------------------------------------------------

# Create temp WaterSourceNativeID dataframe of unique water source.
def assignWaterSourceNativeID(colrowValue):
    string1 = str(colrowValue)
    outstring = "WaDEUSBR_WS" + string1
    return outstring

dfWaterSourceNativeID = pd.DataFrame()
dfWaterSourceNativeID['in_WaterSourceType'] = dfout['WaterSource']
dfWaterSourceNativeID = dfWaterSourceNativeID.drop_duplicates()

dftemp = pd.DataFrame(index=dfWaterSourceNativeID.index)
dftemp["Count"] = range(1, len(dftemp.index) + 1)
dfWaterSourceNativeID['in_WaterSourceNativeID'] = dftemp.apply(lambda row: assignWaterSourceNativeID(row['Count']), axis=1)

# ----------------------------------------------------------------------------------------------------

# Retreive WaDE Custom water source native ID
def retrieveWaterSourceNativeID(A):
    if (A == '') or (pd.isnull(A)):
        outList = ''
    else:
        ml = dfWaterSourceNativeID.loc[(dfWaterSourceNativeID['in_WaterSourceType'] == A), 'in_WaterSourceNativeID']
        if not (ml.empty):  # check if the series is empty
            outList = ml.iloc[0]
        else:
            outList = ''
    return outList

dfout['in_WaterSourceNativeID'] = dfout.apply(lambda row: retrieveWaterSourceNativeID(row['WaterSource']), axis=1)
dfout

# Creating WaDE Custom reportingunit native ID for easy site identificaiion
# ----------------------------------------------------------------------------------------------------

# Create temp ReportingUnitNativeID dataframe of unique reporting units.
def assignReportingUnitNativeID(colrowValue):
    string1 = str(colrowValue)
    outstring = "WaDEUSBR_RU" + string1
    return outstring

dfReportingUnitNativeID = pd.DataFrame()
dfReportingUnitNativeID['in_ReportingUnitName'] = dfout['in_ReportingUnitName']
dfReportingUnitNativeID = dfReportingUnitNativeID.drop_duplicates()

dftemp = pd.DataFrame(index=dfReportingUnitNativeID.index)
dftemp["Count"] = range(1, len(dftemp.index) + 1)
dfReportingUnitNativeID['in_ReportingUnitNativeID'] = dftemp.apply(lambda row: assignReportingUnitNativeID(row['Count']), axis=1)

# ----------------------------------------------------------------------------------------------------

# Retreive WaDE Custom reportingunit native ID
def retrieveReportingUnitNativeID(A):
    if (A == '') or (pd.isnull(A)):
        outList = ''
    else:
        ml = dfReportingUnitNativeID.loc[(dfReportingUnitNativeID['in_ReportingUnitName'] == A), 'in_ReportingUnitNativeID']
        if not (ml.empty):  # check if the series is empty
            outList = ml.iloc[0]
        else:
            outList = ''
    return outList

dfout['in_ReportingUnitNativeID'] = dfout.apply(lambda row: retrieveReportingUnitNativeID( row['in_ReportingUnitName']), axis=1)
dfout

# Shapefile Data (to crate WKT geometry input)

#Shapefile input
shfile = gpd.read_file('C:/Users/rjame/Documents/WSWC Documents/MappingStatesDataToWaDE2.0/USBR/Colorado River Basin Natural Flow and Salt Data/Summary Sheet/AggregatedAmounts/RawInputData/byState/ByState.shp')
dfshape = pd.DataFrame(shfile)
dfshape.head(3)

#Create output shape dataframe
dfshapeout = pd.DataFrame(columns = ['STATE', "geometry"])
dfshapeout['STATE'] = dfshape['States']
dfshapeout['geometry'] = dfshape['geometry']
dfshapeout

# Inspect Output Data & Export

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(dfout.dtypes)

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(dfshapeout.dtypes)

# dfout out to CSV.
dfout.to_csv('P_usbrAggMaster.csv', index=False) # The output.
dfshapeout.to_csv('P_usbrGeometry.csv', index=False) # The output geometry.
